+++ _includes/dependencies-blog.html
@@ -16,5 +16,5 @@
 </script>
 <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.1/MathJax.js?config=TeX-AMS_SVG"></script>
 <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/cookieconsent2/3.0.3/cookieconsent.min.js"></script>
-<script type="text/javascript" src="/assets/js/index.blog.min.js?rev=55a663e8e16b11472ffa93c20391c6b7"></script>
+<script type="text/javascript" src="/assets/js/index.blog.min.js?rev=75682930a8b76a447cf36369e8a24756"></script>
 {%include service-worker.html %}
