+++ _includes/dependencies-home.html
@@ -1,3 +1,3 @@
 <script type="text/javascript" src="//cdnjs.cloudflare.com/ajax/libs/cookieconsent2/3.0.3/cookieconsent.min.js"></script>
-<script type="text/javascript" src="/assets/js/index.home.min.js?rev=c8f0037d7af7b41cb855266f404363de"></script>
+<script type="text/javascript" src="/assets/js/index.home.min.js?rev=942453ea69bd6dd93127a4eb140ab49d"></script>
 {%include service-worker.html %}
