+++ _includes/footer.html
@@ -24,6 +24,6 @@
 		{% include social-contacts.html icon-class="blog-footer-icon" break=false %}
 	</div>
 	<div class="d-flex justify-content-center copyright">
-		<h5>Made with <span id="heart" class="heart">♥</span> by {{ site.author }} 'Chicio'</h5>
+		<h5>Made with <span id="heart" class="heart">💝</span> by {{ site.author }} 'Chicio'</h5>
 	</div>
 </footer>
