+++ _includes/lazy-image.html
@@ -1,0 +1,5 @@
+<img id="#{{include.id}}"
+     alt="{{include.alt}}"
+     src=""
+		 data-src="{{include.src}}"
+		 class="{{include.class}} lazy">
