+++ _layouts/home.html
@@ -7,9 +7,10 @@
             <div id="rendering-surface" class="container-fluid"></div>
             <div class="row vertical-center-row text-center">
                 <div id="center-content" class="center-content">
-                    <img src="/assets/images/fabrizio-duroni.jpg"
+                    <img src=""
+                         data-src="/assets/images/fabrizio-duroni.jpg"
                          id="profile-image"
-                         class="img-responsive img-circle"
+                         class="img-responsive img-circle lazy"
                          alt="Fabrizio Duroni"/>
                     <h1 id="profile-name">{{ site.author }}</h1>
                     <h3 id="profile-job" class="profile-job">Software Developer</h3>
