+++ package.json
@@ -65,7 +65,6 @@
     "gsap": "2.0.2",
     "intersection-observer": "0.5.1",
     "orbit-controls-es6": "2.0.0",
-    "scrollmagic": "2.0.5",
     "three": "0.100.0",
     "three-ply-loader": "1.0.0",
     "webfontloader": "1.6.28"
